Display Key moved to Fragment
