+++ OpenKeychain/src/main/res/layout/multi_user_ids_fragment.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<org.sufficientlysecure.keychain.ui.widget.FixedListView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/view_key_user_ids"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content" />
